Add signup and login, look at description
Through Django's abstract classes and interfaces I redefined the user class again to be able to make use of the methods and functionalities that are already implemented in Django.

Now the home page shows the user's name.

Through Django's base authentication and user creation forms I re-wrote CreateNewUser and Login.

From the views class I made the signUp and signIn methods again, in addition the home method requires the user to be logged in to access.

Finally, in the settings file I changed the language to Spanish so that it is the language in which Django returns form errors and I modified the value of AUTH_USER_MODEL to tell Django where the user I am working with is located.