Add csv generation for reports, see description
CSV generation now works for applicants and announcements, the story will be on hold because the scholarship remake is still on going. I also add a logo and tittles for all the pages that I had worked on before. I refactored the role assign functionality. And the assets folder was deleted